requirements.txt
# Framework para la interfaz web
streamlit==1.55.0
plotly==6.6.0

# SDK de Google Cloud para IA y BigQuery
google-cloud-aiplatform>=1.47.0
google-cloud-bigquery>=3.19.0

# Manejo de datos y optimización de consultas
pandas>=2.2.1
pyarrow==23.0.1
db-dtypes>=1.2.0

# Utilidades
protobuf>=3.20
python-dotenv==1.2.2
pytest==9.0.2